+++ src/shared.cpp
@@ -4,17 +4,13 @@
 bool validKummerParameters(double a, double b, bool warn) {
   if(a < 0) {
     if(warn) {
-      char msg[55];
-      std::sprintf(msg, "Parameter a is less than zero: %f", a);
-      Rcpp::warning(msg);
+      Rcpp::warning("Parameter a is less than zero: %f\n", a);
     }
     return false;
   }
   if(b < a) {
     if(warn) {
-      char msg[256];
-      std::sprintf(msg, "Wrong parameters: b cannot be less than a: %f < %f", b, a);
-      Rcpp::warning(msg);
+      Rcpp::warning("Wrong parameters: b cannot be less than a: %f < %f\n", b, a);
     }
     return false;
   }
@@ -26,9 +22,7 @@ bool isInteger(double x, bool warn) {
     return false;
   if (((x < 0.0) ? std::ceil(x) : std::floor(x)) != x) {
     if (warn) {
-      char msg[55];
-      std::sprintf(msg, "non-integer: %f", x);
-      Rcpp::warning(msg);
+      Rcpp::warning("Non-integer: %f\n", x);
     }
     return false;
   }
@@ -40,9 +34,7 @@ bool validProbability(double p, bool warn) {
     return true;
   } else {
     if(warn) {
-      char msg[55];
-      std::sprintf(msg, "invalid probability: %f", p);
-      Rcpp::warning(msg);
+      Rcpp::warning("Invalid probability: %f\n", p);
     }
     return false;
   }
@@ -51,7 +43,7 @@ bool validProbability(double p, bool warn) {
 bool isInadmissible(double x, bool warn) {
   if(Rcpp::NumericVector::is_na(x) || Rcpp::traits::is_nan<REALSXP>(x)) {
     if(warn)
-      Rcpp::warning("NA/NaNs given in input");
+      Rcpp::warning("NA/NaNs given in input\n");
     return true;
   } else {
     return false;
@@ -61,7 +53,7 @@ bool isInadmissible(double x, bool warn) {
 bool validPbParameters(double alpha, double beta, double c, bool warn) {
   if(alpha > 0 && beta > 0 && c > 0) {
     if(warn) {
-      Rcpp::warning("negative parameters for mpb");
+      Rcpp::warning("Negative parameters for mpb\n");
     }
     return true;
   } else {
@@ -70,7 +62,5 @@ bool validPbParameters(double alpha, double beta, double c, bool warn) {
 }
 
 void reportGslError(int status) {
-  char msg[55];
-  std::sprintf(msg, "GSL Error #%d occured", status);
-  Rcpp::warning(msg);
+  Rcpp::warning("GSL Error #%d occured\n", status);
 }