+++ NEWS.md
@@ -1,6 +1,10 @@
 # Todo
 * Add autoconf script to set user-built mpfr library
 
+# scModels 1.0.4 2023-01-23
+* Fix warnings with sprintf
+* Update travis
+
 # scModels 1.0.3 2022-03-17
 * Update `mpreal.h` version 3.6.2 -> 3.6.9